*** First Release Version: beta 0.1 ***
This is a remake of the previous SangoKushi1 game, This new version will
completely abandon the WindowsMaker auto generated codes and all the GUI
code will now be done by hand. New Design patterns and frameworks are
implemented for the 2nd Sangokushi game, Game data are no longer hard
coded in the source code, instead it will be parsed from an Excel (.xsl)
file.

Features:
* Created the stage for Start Game.
* Added in an Excel parser.
* Added in some basic models.